linux: fix "model name" display in `/proc/cpuinfo` for aarch64 devices
The restriction of displaying model name, called "compat", is used for 32 bit
tasks only, as some old ARMv5/6 devices may do not have a vaild `midr` passing
to cpuinfo, and it's totally useless for ARMv7 and later.

So, it's fine to remove the restriction for ARMv8 devices, and then we can see
"ARMv8 Processor rev X (v8l)" displayed on our LuCI index instead of "?".

Runtime-tested on: bcm27xx, rockchip

Signed-off-by: Tianling Shen <cnsztl@gmail.com>